+++ app/src/views/ElementView.vue
@@ -149,9 +149,16 @@ watch(
     />
 
     <!-- Show the main body and fill the remaining screen space -->
-    <div class="mt-4 flex-1 flex-grow overflow-y-scroll px-8">
+    <div
+      class="mt-4 flex h-[90%] flex-grow flex-col justify-between overflow-y-scroll px-8"
+    >
       <BlockPage v-if="element?.type == 'blk'" :block="element" />
       <RegPage v-else-if="element?.type == 'reg'" :reg="element" />
+
+      <!-- Display footer text if available -->
+      <div v-if="store.footerText" class="mx-auto mb-16 max-w-[800px] text-xs">
+        {{ store.footerText }}
+      </div>
     </div>
   </div>
 </template>